Stage 1 Restructure
Stage 1 of restructure.
- Set up folders
- Moves files into appropriate folders
- Updated imports to restructure
- Updated pyproject
