+++ ImageNexusQt.pyproject
@@ -1,9 +1,9 @@
 {
     "files": [
         "imagenexus.py",
-        "form.ui",
-        "resources.qrc",
-        "about.ui",
-        "aboutDialog.py"
+        "src/ui/form.ui",
+        "resources/resources.qrc",
+        "src/ui/about.ui",
+        "src/utils/aboutDialog.py"
     ]
 }
