+++ resources/rc_resources.py
