+++ src/core/pixelizer.py
