+++ src/core/qr_generator.py
